requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.0
discord.py==1.7.3
et-xmlfile==1.1.0
fake-useragent==0.1.11
h11==0.12.0
httmock==1.4.0
http-request-randomizer==1.3.2
idna==2.10
loguru==0.5.3
lxml==4.6.5
multidict==5.2.0
nordvpn-connect==0.0.7
nordvpn-switcher==0.2.9
numpy==1.21.4
openpyxl==3.0.9
outcome==1.1.0
pandas==1.3.4
pathlib==1.0.1
psutil==5.7.2
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
random-user-agent==1.0.1
requests==2.24.0
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.25.11
webdriver-manager==3.5.2
wsproto==1.0.0
yarl==1.7.2
